+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added support for multi touch pointer events.
 - Improved performance by only updating the `BitmapSprite` offset when it's visible on screen.
+- Improved the handling of swipe actions.
 - Fixed a bug in the handling of nodes, ensuring a name is always provided.
 - Hide debug elements for physics bodies by default.